refactor: change email service implementation
- Updated EmailService to use IEmailSender interface, implemented by SmtpEmailSender.
- Adjusted ServiceCollectionExtensions to register the new email sender.
- Refactored namespaces for better organization and adherence to dependency injection best practices.
