+++ NojectServer/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.Options;
 using NojectServer.Configurations;
-﻿using NojectServer.Middlewares;
+using NojectServer.Middlewares;
 using NojectServer.Services.Auth.Implementations;
 using NojectServer.Services.Auth.Interfaces;
 using NojectServer.Services.Auth.Validation.Implementations;
@@ -9,7 +9,8 @@ using NojectServer.Services.Collaborators.Implementations;
 using NojectServer.Services.Collaborators.Interfaces;
 using NojectServer.Services.Common.Implementations;
 using NojectServer.Services.Common.Interfaces;
-using NojectServer.Services.Email;
+using NojectServer.Services.Email.Implementations;
+using NojectServer.Services.Email.Interfaces;
 
 namespace NojectServer.DependencyInjection;
 
@@ -19,6 +20,7 @@ public static class ServiceCollectionExtensions
     {
         // Register all services
         services.AddScoped<IAuthService, AuthService>();
+        services.AddTransient<IEmailSender, SmtpEmailSender>();
         services.AddTransient<IEmailService, EmailService>();
         services.AddSingleton<IPasswordService, PasswordService>();
         services.AddSingleton<ITokenService, TokenService>();
