+++ NojectServer/Services/Email/Interfaces/IEmailSender.cs
@@ -1,0 +1,8 @@
+﻿using MimeKit;
+
+namespace NojectServer.Services.Email.Interfaces;
+
+public interface IEmailSender
+{
+    Task SendAsync(MimeMessage message);
+}
