+++ NojectServer/Services/Email/Interfaces/IEmailService.cs
@@ -1,7 +1,7 @@
 ﻿using NojectServer.Models;
 using Task = System.Threading.Tasks.Task;
 
-namespace NojectServer.Services.Email;
+namespace NojectServer.Services.Email.Interfaces;
 
 public interface IEmailService
 {
